requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
drf-yasg==1.20.0
flake8==3.8.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==20.8
pkg-resources==0.0.0
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==2.0.1
pyparsing==2.4.7
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
